warning reduction / grooming

+++ util/CheckSums.h
@@ -55,7 +55,7 @@ namespace CheckSums {
     constexpr inline void CheckSumCombine(uint32_t& sum, std::string_view sv) noexcept {
         for (auto t : sv)
             CheckSumCombine(sum, t);
-        sum += sv.size();
+        sum += static_cast<uint32_t>(sv.size());
         sum %= CHECKSUM_MODULUS;
     }
     constexpr inline void CheckSumCombine(uint32_t& sum, const char* s) noexcept
